backend/product-service/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.5.5
